rename: spec 폴더 이동 #16
