+++ tests/unit/review/short-review.repository.spec.ts
@@ -1,8 +1,8 @@
 import { AttractionElement } from '@prisma/client';
 import { ShortReviewRepository } from '../../../src/domain/review/repository/short-review.repository';
 import { Test } from '@nestjs/testing';
-import { PrismaModule } from '../../../src/global/config/prisma/prisma.module';
 import { IShortReview } from '../../../src/domain/review/reviews.interface';
+import { PrismaModule } from '../../../src/global/database/prisma/prisma.module';
 
 describe('ShortReviewRepository', () => {
   let repository: ShortReviewRepository;
